restructure benchmark for regex
