+++ 2023/regex/benchmark/timer.py
@@ -6,6 +6,6 @@ def timer_decorator(func: Callable[..., Any]) -> Callable[...,Any]:
         start_time = time.time()
         result = func(*args, **kwargs)
         end_time = time.time()
-        print(f"{func.__name__} took {end_time - start_time:.5f} seconds to execute.")
+        print(f"{func.__name__} with input {args[-1]} took {end_time - start_time:.5f} seconds to execute.")
         return result
     return wrapper
